Inline some functions to eliminate confusion
